separated Systems to a submodule (#157)
* separated Systems to a submodule

* separated utils in a submodule
