+++ wannierberri/__init__.py
@@ -21,12 +21,11 @@ from .__main import integrate, tabulate, integrate_options, tabulate_options, we
 from . import symmetry
 from .__tabulate import TABresult
 from .__grid import Grid
-from .__system import System
-from .__system_w90 import System_w90
-from .__system_fplo import System_fplo
-from .__system_tb import System_tb
-from .__system_tb_py import System_PythTB, System_TBmodels
-from .__system_ASE import System_ASE
+from .system.system_w90 import System_w90
+from .system.system_fplo import System_fplo
+from .system.system_tb import System_tb
+from .system.system_tb_py import System_PythTB, System_TBmodels
+from .system.system_ASE import System_ASE
 from .__path import Path
 from .__parallel import Parallel
 from . import calculators
