+++ wannierberri/system/system_ASE.py
@@ -14,8 +14,8 @@
 
 import numpy as np
 import multiprocessing
-from .__utility import real_recip_lattice
-from .__system_w90 import System_w90
+from wannierberri.__utility import real_recip_lattice
+from .system_w90 import System_w90
 from termcolor import cprint
