+++ wannierberri/system/system_w90.py
@@ -15,9 +15,9 @@
 import numpy as np
 import functools
 import multiprocessing
-from .__utility import iterate3dpm, real_recip_lattice, fourier_q_to_R
-from .__system import System
-from .__w90_files import EIG, MMN, CheckPoint, SPN, UHU, SIU, SHU
+from wannierberri.__utility import iterate3dpm, real_recip_lattice, fourier_q_to_R
+from .system import System
+from wannierberri.__w90_files import EIG, MMN, CheckPoint, SPN, UHU, SIU, SHU
 from time import time
