+++ wannierberri/utils/tab_plot.py
@@ -27,11 +27,11 @@ Tested
         Usage example:
                 line: ::
 
-                    python3 -m wannierberri.tab_plot tab_result.pickle type=Line quantity=True kpath=0,0,0,0,0,40 namelist=G,Z qtype=berry component=z
+                    python3 -m wannierberri.utils.tab_plot tab_result.pickle type=Line quantity=True kpath=0,0,0,0,0,40 namelist=G,Z qtype=berry component=z
 
                 plane: ::
 
-                    python3 -m wannierberri.tab_plot tab_result.pickle type=Plane quantity=True Efermi=-0.5 vec1=1,0,0 vec2=0,1,0 qtype=berry component=z
+                    python3 -m wannierberri.utils.tab_plot tab_result.pickle type=Plane quantity=True Efermi=-0.5 vec1=1,0,0 vec2=0,1,0 qtype=berry component=z
 
         Options
             -h
